+++ user-service/src/main/java/com/bank/user/repository/CustomerRepository.java
@@ -6,7 +6,7 @@ import org.springframework.stereotype.Repository;
 import com.bank.user.entity.CustomerEntity;
 
 @Repository
-public interface CustomerRepository extends JpaRepository<CustomerEntity, Long> {
+public interface CustomerRepository extends JpaRepository<CustomerEntity, Integer> {
 
 	CustomerEntity findCustomerByPhoneNo(String phoneNo);
 }
